+++ src/main/java/com/microsoft/graph/CoreConstants.java
@@ -1,6 +1,9 @@
 package com.microsoft.graph;
 
+import javax.annotation.Nonnull;
+import java.util.HashMap;
 import java.util.Locale;
+import java.util.Map;
 
 /**
  * Core Constants for use in other classes.
@@ -11,7 +14,7 @@ public final class CoreConstants {
     private static class VersionValues {
         private static final int MAJOR = 3;
         private static final int MINOR = 0;
-        private static final int PATCH = 6;
+        private static final int PATCH = 7;
     }
 
     /**
@@ -105,10 +108,40 @@ public final class CoreConstants {
      * Collection Response Method Name Constants
      */
     public static class CollectionResponseMethods {
-        private CollectionResponseMethods(){}
-        /** Method name constant for getOdataDeltaLink in collection responses*/
+        private CollectionResponseMethods() {}
+        /** Method name constant for getOdataDeltaLink in collection responses */
         public static final String GET_ODATA_DELTA_LINK = "getOdataDeltaLink";
-        /** Method name constant for getOdataNextLink in collection responses*/
+        /** Method name constant for getOdataNextLink in collection responses */
         public static final String GET_ODATA_NEXT_LINK = "getOdataNextLink";
     }
+
+    /**
+     * Url Replacement Constants
+     * Example: /users/{id} -> /me
+     */
+    public static class ReplacementConstants {
+        private ReplacementConstants() {
+        }
+        /** Default token to replace the id value in 'user/{id}' */
+        public static final String USER_ID_TOKEN_TO_REPLACE = "TokenToReplace";
+        /** Default endpoint to be replaced by /me
+         * users/TokenToReplace -> /me
+         */
+        public static final String USERS_ENDPOINT_WITH_REPLACE_TOKEN = "/users/" + USER_ID_TOKEN_TO_REPLACE;
+        /** /me endpoint constant */
+        public static final String ME_ENDPOINT = "/me";
+
+        /**
+         * The default replacement pairs for /users/{id} -> /me
+         * @return the default replacement pairs
+         */
+        @Nonnull
+        public static Map<String, String> getDefaultReplacementPairs() {
+            HashMap<String, String> defaultReplacementPairs = new HashMap<>();
+            defaultReplacementPairs.put(USERS_ENDPOINT_WITH_REPLACE_TOKEN, ME_ENDPOINT);
+            return defaultReplacementPairs;
+        }
+
+    }
+
 }