+++ src/main/java/com/microsoft/graph/requests/GraphClientFactory.java
@@ -1,7 +1,11 @@
 package com.microsoft.graph.requests;
 
+import com.microsoft.graph.CoreConstants;
 import com.microsoft.graph.requests.middleware.GraphTelemetryHandler;
+import com.microsoft.graph.requests.options.GraphClientOption;
 import com.microsoft.kiota.http.KiotaClientFactory;
+import com.microsoft.kiota.http.middleware.UrlReplaceHandler;
+import com.microsoft.kiota.http.middleware.options.UrlReplaceHandlerOption;
 import okhttp3.Interceptor;
 import okhttp3.OkHttpClient;
 
@@ -77,6 +81,8 @@ public class GraphClientFactory {
     public static Interceptor[] createDefaultGraphInterceptors(@Nonnull GraphClientOption graphClientOption) {
         List<Interceptor> handlers = new ArrayList<>();
         addDefaultFeatureUsages(graphClientOption);
+
+        handlers.add(new UrlReplaceHandler(new UrlReplaceHandlerOption(CoreConstants.ReplacementConstants.getDefaultReplacementPairs())));
         handlers.add(new GraphTelemetryHandler(graphClientOption));
         handlers.addAll(Arrays.asList(KiotaClientFactory.createDefaultInterceptors()));
         return handlers.toArray(new Interceptor[0]);
@@ -85,5 +91,7 @@ public class GraphClientFactory {
     private static void addDefaultFeatureUsages(GraphClientOption graphClientOption) {
         graphClientOption.featureTracker.setFeatureUsage(FeatureFlag.RETRY_HANDLER_FLAG);
         graphClientOption.featureTracker.setFeatureUsage(FeatureFlag.REDIRECT_HANDLER_FLAG);
+        graphClientOption.featureTracker.setFeatureUsage(FeatureFlag.URL_REPLACEMENT_FLAG);
+        graphClientOption.featureTracker.setFeatureUsage(FeatureFlag.BATCH_REQUEST_FLAG);
     }
 }